Make `SummaryNumber` rendering more readable,
Create specialized `MetricNumber` component to
hide the majority of pre-2021Q4 logic.

Addresses https://github.com/woocommerce/google-listings-and-ads/pull/109#discussion_r557311353.
