+++ js/src/reports/index.js
@@ -2,14 +2,7 @@
  * External dependencies
  */
 import { __ } from '@wordpress/i18n';
-import {
-	SummaryList,
-	SummaryNumber,
-	Chart,
-	FilterPicker,
-} from '@woocommerce/components';
-import { Tooltip } from '@wordpress/components';
-import GridiconInfoOutline from 'gridicons/dist/info-outline';
+import { SummaryList, Chart, FilterPicker } from '@woocommerce/components';
 import { getQuery } from '@woocommerce/navigation';
 
 /**
@@ -18,8 +11,8 @@ import { getQuery } from '@woocommerce/navigation';
 import TabNav from '../tab-nav';
 import AppDateRangeFilterPicker from '../dashboard/app-date-range-filter-picker';
 import CompareProgramsTableCard from './compare-programs-table-card';
+import MetricNumber from './metric-number';
 import '../dashboard/index.scss';
-import './index.scss';
 import { programsFilterConfig } from './filter-config';
 import metricsData from './mocked-metrics-data'; // Mocked data
 
@@ -28,14 +21,14 @@ import metricsData from './mocked-metrics-data'; // Mocked data
  *
  * @type {Map<string, string>}
  */
-const performanceMetrics = new Map( [
+const performanceMetrics = [
 	[ 'netSales', __( 'Net Sales', 'google-listings-and-ads' ) ],
 	[ 'itemsSold', __( 'Items Sold', 'google-listings-and-ads' ) ],
 	[ 'conversions', __( 'Conversions', 'google-listings-and-ads' ) ],
 	[ 'clicks', __( 'Clicks', 'google-listings-and-ads' ) ],
 	[ 'impressions', __( 'Impressions', 'google-listings-and-ads' ) ],
 	[ 'totalSpend', __( 'Total Spend', 'google-listings-and-ads' ) ],
-] );
+];
 
 const ProgramsReports = () => {
 	// TODO: this data should come from backend API.
@@ -93,6 +86,12 @@ const ProgramsReports = () => {
 		},
 	];
 
+	// Show only available data.
+	// Until ~Q4 2021, free listings, may not have all metrics.
+	const availableMetrics = performanceMetrics.filter(
+		( [ key ] ) => data[ key ]
+	);
+
 	const trackEventReportId = 'reports-programs';
 
 	return (
@@ -109,49 +108,15 @@ const ProgramsReports = () => {
 			</div>
 			<div className="gla-reports__performance">
 				<SummaryList>
-					{ () => {
-						return Array.from(
-							performanceMetrics,
-							function renderSingleMetric( [ key, label ] ) {
-								// Show only available data.
-								// Until ~Q4 2021, free listings, may not have all metrics.
-								if ( ! data[ key ] ) {
-									return undefined;
-								}
-								let markedLabel = label;
-								// Until ~Q4 2021, metrics for all programs, may lack data for free listings.
-								if ( data[ key ].missingFreeListingsData ) {
-									const infoText = __(
-										'This data is available for paid campaigns only.',
-										'google-listings-and-ads'
-									);
-									markedLabel = (
-										<div className="gla-reports__metric-label">
-											{ label }
-											<Tooltip text={ infoText }>
-												<span>
-													<GridiconInfoOutline
-														className="gla-reports__metric-infoicon"
-														role="img"
-														aria-label={ infoText }
-														size={ 16 }
-													/>
-												</span>
-											</Tooltip>
-										</div>
-									);
-								}
-								return (
-									<SummaryNumber
-										key={ key }
-										label={ markedLabel }
-										value={ data[ key ].value }
-										delta={ data[ key ].delta }
-									/>
-								);
-							}
-						).filter( Boolean ); // Ignore undefined elements.
-					} }
+					{ () =>
+						availableMetrics.map( ( [ key, label ] ) => (
+							<MetricNumber
+								key={ key }
+								label={ label }
+								data={ data[ key ] }
+							/>
+						) )
+					}
 				</SummaryList>
 				<Chart
 					data={ chartData }
