+++ js/src/reports/metric-number.scss
